Move SqliteMemoryManager::MallocArena into separate compilation unit
